+++ translate/providers/__init__.py
@@ -1,0 +1,7 @@
+#!/usr/bin/env python
+# encoding: utf-8
+
+from .mymemory_translated import MyMemoryProvider  # noqa
+from .microsoft import MicrosoftProvider  # noqa
+
+__all__ = ['MyMemoryProvider', 'MicrosoftProvider']